pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "og-backend"
description = "Open Graph Backend"
version = "0.1.0"
readme = "README.md"
authors = ["og-backend.app"]


[tool.poetry.dependencies]
python = "3.8.9"

django = "^2.2.19"
django-split-settings = "^1.0"
django-axes = "^5.14"
django-csp = "^3.7"
django-health-check = "^3.16"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.0"
django-stubs-ext = "^0.2"

psycopg2-binary = "^2.8"
gunicorn = "^20.0"
python-decouple = "^3.4"
bcrypt = "^3.2"
structlog = "^21.1"
djangorestframework = "^3.12.4"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2"
django-querycount = "^0.7"
django-migration-linter = "^2.5"
django-extra-checks = "^0.11"
django-coverage-plugin = "^1.8"
nplusone = "^1.0"

wemake-python-styleguide = "^0.15"
flake8-pytest-style = "^1.4"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"
nitpick = "^0.26"

pytest = "^6.2"
pytest-django = "^4.2"
pytest-cov = "^2.11"
pytest-randomly = "^3.7"
pytest-deadfixtures = "^2.2"
pytest-testmon = "^1.1"
pytest-timeout = "^1.4"
django-test-migrations = "^1.1"
hypothesis = "^6.10"

mypy = "^0.812"
django-stubs = "^1.8"

sphinx = "^3.5"
sphinx-autodoc-typehints = "^1.12"
tomlkit = "^0.7"
doc8 = "^0.8"

yamllint = "^1.26"
safety = "^1.10"
dotenv-linter = "^0.2"
polint = "^0.4"
dennis = "^0.9"
dump-env = "^1.3"
ipython = "^7.21"
